[DONE] refactor name of pages and templates
