+++ src/D3vS1m.WebUI/ClientApp/router/routes.js
@@ -1,4 +1,4 @@
-import HomePage from '../components/pages/home-page'
+import HomePage from '../components/pages/home'
 
 export const routes = [
   { name: 'home', path: '/', component: HomePage, display: 'Home', icon: 'home' }
